requirements.txt
pytest==8.4.0
requests==2.32.4
Faker==37.3.0
pytest-reporter-html1==0.9.3
python-dotenv==1.1.0